moved remark extensions to index
