--- src/markdown/userlink/remark-userlink.js
@@ -1,11 +1,0 @@
-import syntax from './micromark-extension-userlink';
-import { fromMarkdown } from './mdast-extension-userlink';
-import { add } from '../utils';
-
-export default username;
-
-function username() {
-  const data = this.data();
-  add(data, 'micromarkExtensions', syntax);
-  add(data, 'fromMarkdownExtensions', fromMarkdown);
-}
